kibi-select now returns the full query object instead of the query ID only.
This allows to get access to metadata about the query, without
having to re-fetch it again. #1305
Corrected some issues with the kibi-select when automatically selecting
items.

Removed the refresh button in templates/queries settings sections #686

Bumped version of angular from 1.4.7 to 1.4.8 in order to get the fix
for ngIf on the empty option. https://github.com/angular/angular.js/pull/13029